// portalWeb/src/app/models/reportes/control_asistencia.ts
export interface ControlAsistencia {
  colaborador:    string;
  identificacion: string;
  codBiometrico:  string;
  udn:            string;
  area:           string;
  subCentroCosto: string;
  fecha:          Date|string;
  //laboral
  idTurno?:                 string;
  entrada:                  Date|string;
  idMarcacionEntrada?:      number;
  marcacionEntrada:         Date | string;
  salida:                   Date|string;
  idMarcacionSalida?:       number;
  marcacionSalida:          Date | string;
  totalHoras?:              null | string;
  idSolicitudEntrada?:      null | string;
  idFeatureEntrada?:        null;
  estadoEntrada:            string;
  fechaSolicitudEntrada?:   Date;
  usuarioSolicitudEntrada?: string;
  estadoSolicitudEntrada?:  null;
  idSolicitudSalida?:       null | string;
  idFeatureSalida?:         null;
  estadoSalida:             string;
  fechaSolicitudSalida?:    Date;
  usuarioSolicitudSalida?:  string;
  estadoSolicitudSalida?:   null;
  //receso
  idTurnoR?:                      null | string;
  entradaR:                       Date | string;
  idMarcacionEntradaR?:           number;
  marcacionEntradaR:              string;
  salidaR:                        Date | string;
  idMarcacionSalidaR?:            number;
  marcacionSalidaR:               string;
  totalHorasR?:                   null | string;
  idSolicitudEntradaReceso?:      null;
  idFeatureEntradaReceso?:        null;
  estadoEntradaReceso?:           null | string;
  fechaSolicitudEntradaReceso?:   Date;
  usuarioSolicitudEntradaReceso?: string;
  estadoSolicitudEntradaReceso?:  null;
  idSolicitudSalidaReceso?:       null;
  idFeatureSalidaReceso?:         null;
  estadoSalidaReceso?:            null | string;
  fechaSolicitudSalidaReceso?:    Date;
  usuarioSolicitudSalidaReceso?:  string;
  estadoSolicitudSalidaReceso?:   null;
  // novedades: NovedadesA[];
  // solicitudes: SolicitudesA[];
  idMarcacionN?:      number;
  idSolicitudN?:      string;
  usuarioAprobadorN?: string;
  fechaAprobacionN?:  string;
  descripcionN?:      string;
  minutosNovedadN?:   string;
  estadoMarcacionN?:  string;
}